Added hilt and coroutine  dependency

+++ app/src/main/java/com/anurag/spacexextraaedgetask/model/Height.kt
@@ -8,5 +8,5 @@ import com.google.gson.annotations.SerializedName
 
 data class Height(
     @SerializedName("meters" ) var meters : Double? = null,
-    @SerializedName("feet"   ) var feet   : Int?    = null
+    @SerializedName("feet"   ) var feet   : Double?    = null
 )
